pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "email-otp-api"
version = "1.0.0"
description = "A production-grade Email OTP Verification Microservice"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "email_otp_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic = {version = "^2.5.0", extras = ["email"]}
SQLAlchemy = "^2.0.23"
alembic = "^1.12.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
